+++ src/components/control/multi-select.jsx
@@ -18,9 +18,10 @@ import NoOptions from '../no-options/no-options.jsx';
 import TagList from '../tag-list/tag-list.jsx';
 import Tag from '../tag/tag.jsx';
 import useDropdownClose from '../../hooks/use-dropdown-close.js';
-import useMountedEffect from '../../hooks/use-mounted-effect.js';
-import styles from './multi-select.css';
 import useForwardedRef from '../../hooks/use-forwarded-ref.js';
+import useMountedEffect from '../../hooks/use-mounted-effect.js';
+import useValidation from '../../hooks/use-validation.js';
+import styles from './multi-select.css';
 
 function getFormControlChildrenContainerClassName(readOnly, validationMessage, classNames) {
   if (readOnly) {
@@ -38,7 +39,6 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   const [autocompleteValue, setAutocompleteValue] = useState('');
   const [expanded, setExpanded] = useState(false);
   const selfRef = useForwardedRef(ref);
-  const [validationMessage, setValidationMessage] = useState(props.validationMessage);
   const [value, setValue] = useState(props.value || []);
   const valueRefs = value.map(() => createRef());
   const visibleOptions = getVisibleOptions();
@@ -57,7 +57,6 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     container: props.containerRef,
     root: useRef(),
   };
-
   const classNames = {
     formControlChildrenContainer: styles['form-control-children-container'],
     formControlChildrenContainerInvalid: styles['form-control-children-container-invalid'],
@@ -69,6 +68,8 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
     tagList: styles['tag-list'],
     ...props.classNames,
   };
+
+  const [validationMessage, validate] = useValidation(props.validationMessage, refs.autocomplete);
 
   const dropdownContents = () => {
     if (!expanded) {
@@ -132,14 +133,8 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   function onBlur(event) {
     if (!refs.autocomplete.current) return;
     if (refs.root.current.contains(event.relatedTarget)) return;
-
-    refs.autocomplete.current.setCustomValidity('');
-    if (refs.autocomplete.current.validity.valid) {
-      refs.autocomplete.current.setCustomValidity(props.validationMessage);
-      setValidationMessage(props.validationMessage || '');
-    } else {
-      setValidationMessage(refs.autocomplete.current.validationMessage);
-    }
+    validate();
+    // onDropdownClose();
   }
 
   function onAutocompleteChange(event) {
@@ -203,10 +198,6 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
   }
 
   useDropdownClose(refs.container, expanded, onDropdownClose);
-
-  useEffect(() => {
-    setValidationMessage(props.validationMessage);
-  }, [props.validationMessage]);
 
   useEffect(() => {
     props.onInvalid({ detail: { validationMessage } });